+++ src/lib.rs
@@ -31,3 +31,4 @@ pub mod status;
 pub mod teelog;
 pub mod type_override;
 pub mod unicode_relocation;
+pub mod warn_execute;